Fetch fresh pricing data asynchronously
