+++ app/urls.py
@@ -10,7 +10,7 @@ from wagtail.core import urls as wagtail_urls
 from wagtail.documents import urls as wagtaildocs_urls
 from wagtailautocomplete.urls.admin import urlpatterns as autocomplete_admin_urls
 
-from app.views import ShippingCostView, StripeCustomerPortalView
+from app.views import CartOptionsView, ShippingCostView, StripeCustomerPortalView
 
 # from wagtail_transfer import urls as wagtailtransfer_urls
 
@@ -35,6 +35,7 @@ urlpatterns = [
     path(
         ShippingCostView.url_pattern, ShippingCostView.as_view(), name="shippingcosts"
     ),
+    path(CartOptionsView.url_pattern, CartOptionsView.as_view(), name="cartoptions"),
     # re_path(r'^wagtail-transfer/', include(wagtailtransfer_urls)),
     # For anything not caught by a more specific rule above, hand over to Wagtail's serving mechanism
     re_path(r"", include(wagtail_urls)),
